Rename package timeout to timeoutconn.
