+++ pkg/timeoutconn/timeoutconn_test.go
@@ -1,4 +1,4 @@
-package timeout
+package timeoutconn
 
 import (
 	"bufio"
